+++ myems-api/core/advancedreport.py
@@ -382,18 +382,10 @@ class AdvancedReportRun:
     def on_put(req, resp, id_):
         """Handles PUT requests"""
         admin_control(req)
-        try:
-            raw_json = req.stream.read().decode('utf-8')
-        except Exception:
-            raise falcon.HTTPError(status=falcon.HTTP_400,
-                                   title='API.BAD_REQUEST',
-                                   description='API.FAILED_TO_READ_REQUEST_STREAM')
 
         if not id_.isdigit() or int(id_) <= 0:
             raise falcon.HTTPError(status=falcon.HTTP_400, title='API.BAD_REQUEST',
                                    description='API.INVALID_ADVANCED_REPORT_ID')
-        # ignore the payload
-        new_values = json.loads(raw_json)
 
         cnx = mysql.connector.connect(**config.myems_reporting_db)
         cursor = cnx.cursor()